requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
astroid==3.2.2
attrs==23.2.0
click==8.1.7
coverage==7.5.3
dill==0.3.8
durable-rules==2.0.28
exceptiongroup==1.2.1
iniconfig==2.0.0
isort==5.13.2
libs==0.0.10
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
psycopg2-binary==2.9.9
Pygments==2.18.0
pylint==3.2.3
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
rule-engine==4.4.0
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.30
tomli==2.0.1
tomlkit==0.12.5
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
